requirements.txt
flask==2.3.3
flask-cors==4.0.0
cohere==4.37
speechrecognition==3.10.0
gtts==2.3.2
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3 